+++ Hurrican/src/triggers/Trigger_Spikelift.cpp
@@ -32,13 +32,13 @@ GegnerSpikelift::GegnerSpikelift(int Wert1, int Wert2, bool Light) {
 void GegnerSpikelift::DoDraw() {
     // Unterteil rendern
     //
-    pGegnerGrafix[GegnerArt]->RenderSprite(static_cast<float>(xPos - TileEngine.XOffset),
-                                           static_cast<float>(yPos - TileEngine.YOffset) + 40.0f, 1, 0xFFFFFFFF);
+    pGegnerGrafix[GegnerArt]->RenderSprite(xPos - TileEngine.XOffset,
+                                           yPos - TileEngine.YOffset + 40.0f, 1, 0xFFFFFFFF);
 
     // Oberteil (Spikes) rendern
     if (Value1 == 0)
-        pGegnerGrafix[GegnerArt]->RenderSprite(static_cast<float>(xPos - TileEngine.XOffset),
-                                               static_cast<float>(yPos - TileEngine.YOffset), 0, 0xFFFFFFFF);
+        pGegnerGrafix[GegnerArt]->RenderSprite(xPos - TileEngine.XOffset,
+                                               yPos - TileEngine.YOffset, 0, 0xFFFFFFFF);
 }
 
 // --------------------------------------------------------------------------------------
